add pageIndicator into the header
