+++ src/App.css
@@ -9,7 +9,7 @@
 }
 
 .App-header {
-  background-color: #FFFFFF;
+  background-color: #ffffff;
   min-height: 100vh;
   display: flex;
   flex-direction: column;
@@ -30,76 +30,4 @@
   to {
     transform: rotate(360deg);
   }
-}
-
-.ProgressBar {
-  margin: 0 auto;
-  padding: 2em 0 3em;
-  list-style: none;
-  position: relative;
-  display: flex;
-  justify-content: space-between;
-}
-.ProgressBar-step {
-  text-align: center;
-  position: relative;
-  width: 100%;
-}
-.ProgressBar-step:before,
-.ProgressBar-step:after {
-  content: "";
-  height: 0.5em;
-  background-color: #9f9fa3;
-  position: absolute;
-  z-index: 1;
-  width: 100%;
-  left: -50%;
-  top: 50%;
-  transform: translateY(-50%);
-  transition: all 0.25s ease-out;
-}
-.ProgressBar-step:first-child:before,
-.ProgressBar-step:first-child:after {
-  display: none;
-}
-.ProgressBar-step:after {
-  background-color: #594fd1;
-  width: 0%;
-}
-.ProgressBar-step.is-complete + .ProgressBar-step.is-current:after,
-.ProgressBar-step.is-complete + .ProgressBar-step.is-complete:after {
-  width: 100%;
-}
-.ProgressBar-icon {
-  width: 0.75em;
-  height: 0.75em;
-  background-color: #9f9fa3;
-  fill: #9f9fa3;
-  border-radius: 50%;
-  padding: 0.5em;
-  max-width: 100%;
-  z-index: 10;
-  position: relative;
-  transition: all 0.25s ease-out;
-}
-.is-current .ProgressBar-icon {
-  fill: #594fd1;
-  background-color: #594fd1;
-}
-.is-complete .ProgressBar-icon {
-  fill: #dbf1ff;
-  background-color: #594fd1;
-}
-.ProgressBar-stepLabel {
-  display: block;
-  text-transform: uppercase;
-  color: #9f9fa3;
-  position: absolute;
-  padding-top: 0.5em;
-  width: 100%;
-  transition: all 0.25s ease-out;
-}
-.is-current > .ProgressBar-stepLabel,
-.is-complete > .ProgressBar-stepLabel {
-  color: #594fd1;
 }
